requirements.txt
requests
opencv-python
numpy
tensorflow
paho-mqtt
configparser